+++ project-base/src/Shopsys/ShopBundle/Model/AdvancedSearchOrder/Filter/OrderEmailFilter.php
@@ -51,7 +51,7 @@ class OrderEmailFilter implements AdvancedSearchFilterInterface
             if ($ruleData->value === null || $ruleData->value == '') {
                 $searchValue = '%';
             } else {
-                $searchValue = '%' . DatabaseSearching::getLikeSearchString($ruleData->value) . '%';
+                $searchValue = DatabaseSearching::getFullTextLikeSearchString($ruleData->value);
             }
             $dqlOperator = $this->getContainsDqlOperator($ruleData->operator);
             $parameterName = 'email_' . $index;