+++ project-base/src/Shopsys/ShopBundle/Model/AdvancedSearchOrder/Filter/OrderNumberFilter.php
@@ -54,7 +54,7 @@ class OrderNumberFilter implements AdvancedSearchFilterInterface
                 if ($ruleData->value === null || $ruleData->value == '') {
                     $searchValue = '%';
                 } else {
-                    $searchValue = '%' . DatabaseSearching::getLikeSearchString($ruleData->value) . '%';
+                    $searchValue = DatabaseSearching::getFullTextLikeSearchString($ruleData->value);
                 }
 
                 $dqlOperator = $this->getContainsDqlOperator($ruleData->operator);