+++ project-base/src/Shopsys/ShopBundle/Model/Category/CategoryRepository.php
@@ -433,7 +433,7 @@ class CategoryRepository extends NestedTreeRepository
         $queryBuilder->andWhere(
             'NORMALIZE(ct.name) LIKE NORMALIZE(:searchText)'
         );
-        $queryBuilder->setParameter('searchText', '%' . DatabaseSearching::getLikeSearchString($searchText) . '%');
+        $queryBuilder->setParameter('searchText', DatabaseSearching::getFullTextLikeSearchString($searchText));
     }
 
     /**